Modify: Main 레이아웃 디테일 수정
