+++ src/components/Slide/Slide.scss
@@ -16,7 +16,7 @@
 
   &-wrapper__insides {
     display: flex;
-    padding-top: 2%;
+    padding: 4% 2% 1.5% 0;
     z-index: 1;
     color: #fff;
 
@@ -83,7 +83,8 @@
     .button--prev {
       color: #fff;
       font-size: 20px;
-      &:disabled{
+
+      &:disabled {
         color: #888;
       }
 
@@ -105,7 +106,7 @@
       margin-left: 5px;
       color: #fff;
 
-      &:disabled{
+      &:disabled {
         color: #888;
       }
     }
