--- src/pages/Main/Banner/Banner.scss
@@ -1,29 +1,0 @@
-@import '../../../styles/common.scss';
-
-.banner {
-  @include flex-column-center;
-  height: 383px;
-  padding: 0 0 0 70px;
-  background: center/cover url('https://i.postimg.cc/jjdWxPMZ/background.jpg');
-  color: #292929;
-
-  &__sub-title {
-    font-size: 11px;
-    font-weight: 700;
-  }
-
-  &__title {
-    font-family: 'Noto Sans KR';
-    font-size: 47px;
-  }
-
-  &__content {
-    padding-top: 5px;
-    font-size: 14px;
-    color: #292929;
-
-    p:last-of-type {
-      margin-bottom: 25px;
-    }
-  }
-}
