--- src/pages/Main/Colletions/CollectionsItem/CollectionsItem.scss
@@ -1,52 +1,0 @@
-@import "../../../../styles/common.scss";
-
-.collections-item {
-  position: relative;
-  padding: 2%;
-  width: 48%;
-
-  @media screen and (max-width: 767px) {
-    width: 100%;
-    padding: 5% 0 0 0;
-  }
-
-  &__image {
-    >img {
-      width: 100%;
-    }
-  }
-
-  &__content {
-    @include flex-column-end;
-    align-items: flex-start;
-    position: absolute;
-    top: 0;
-    left: 0;
-    padding: 0 0 15% 12%;
-    height: 100%;
-    color: #fff;
-
-    @media screen and (max-width: 1190px) {
-      padding-left: 10%;
-    }
-
-    @media screen and (max-width: 767px) {
-      padding-bottom: 10%;
-    }
-  }
-
-  &__title {
-    font-size: 24px;
-    font-weight: 600;
-    color: #fff;
-
-
-    @media screen and (max-width: 1190px) {
-      font-size: 22px;
-    }
-
-    @media screen and (max-width: 767px) {
-      font-size: 19px;
-    }
-  }
-}
