--- src/pages/Main/Cover/Cover.scss
@@ -1,36 +1,0 @@
-.cover {
-  position: relative;
-
-  &__video {
-
-
-    >video {
-      width: 100%;
-      height: calc(100vh * 0.8);
-      object-fit: cover;
-    }
-  }
-
-  &__content {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-end;
-    position: absolute;
-    bottom: 15%;
-    left: 5%;
-    color: #fff
-  }
-
-  &__text {}
-
-  &__title {
-    font-size: 40px;
-    font-weight: 600;
-    font-family: Arial, Helvetica, sans-serif;
-
-    @media screen and(max-width: 768px) {
-      font-size: 38px;
-    }
-
-  }
-}
