--- src/pages/Main/Insides/Insides.scss
@@ -1,5 +1,0 @@
-
-.insides__image-list {
-  display: flex;
-  transition: transform 0.2s ease-out 0s;
-}
