+++ src/pages/Main/Main.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import Cover from './Cover/Cover';
-import Products from './Products/Products';
-import Banner from './Banner/Banner';
-import Collaboration from './Collaboration/Collaboration';
-import Colletions from './Colletions/Colletions';
-import Insides from './Insides/Insides';
+import Cover from './components/Cover/Cover';
+import Products from './components/Products/Products';
+import Banner from './components/Banner/Banner';
+import Collaboration from './components/Collaboration/Collaboration';
+import Colletions from './components//Colletions/Colletions';
+import Insides from './components//Insides/Insides';
 
 function Main() {
   return (
