--- src/pages/Main/Products/Products.scss
@@ -1,9 +1,0 @@
-@import "../../../styles/common.scss";
-
-.products {
-  &__image-list {
-    display: flex;
-    width: max-content;
-    transition: transform 0.2s ease-out 0s;
-  }
-}
