+++ src/pages/Main/components/Collaboration/Collaboration.scss
@@ -1,0 +1,69 @@
+@import "../../../../styles/common.scss";
+
+.collaboration {
+  display: flex;
+  margin-top: 100px;
+
+  @media screen and (max-width: 768px) {
+    flex-direction: column;
+  }
+
+  &__image {
+    flex: 1 1 55%;
+
+    >img {
+      width: 100%;
+
+      @media screen and (max-width: 768px) {
+        height: 100%;
+        width: unset;
+      }
+    }
+
+    @media screen and (max-width: 768px) {
+      display: flex;
+      flex-basis: auto;
+      justify-content: center;
+      width: 100%;
+      height: 100vw;
+    }
+  }
+
+  &__content {
+    display: flex;
+    flex: 1 1 45%;
+    justify-content: center;
+    align-items: center;
+    padding-right: 40px;
+    color: #545454;
+
+    @media screen and (max-width: 768px) {
+      flex-basis: auto;
+      padding: calc(100vw / 12) calc(100vw / 25);
+    }
+  }
+
+  &__text {
+    line-height: 25px;
+    width: 55%;
+
+    @media screen and (max-width: 768px) {
+      width: 100%;
+    }
+  }
+
+  &__sub-title {
+    letter-spacing: 1.5px;
+    line-height: 20px;
+    font-size: 10px;
+    font-weight: 600;
+    color: #767676;
+  }
+
+  &__title {
+    line-height: 36px;
+    font-size: 26px;
+    font-weight: 600;
+    color: #000;
+  }
+}
