+++ src/pages/Main/components/Colletions/Colletions.scss
@@ -1,0 +1,32 @@
+.collections {
+  &__title {
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    height: 115px;
+    padding-left: 4%;
+    letter-spacing: 0.5px;
+    font-size: 26px;
+    font-family: 'Inter',
+      sans-serif;
+
+    @media screen and (max-width: 1190px) {
+      font-size: 22px;
+      padding: 0 0 0 4%;
+    }
+
+    @media screen and (max-width: 768px) {
+      padding: 0 0 0 6%;
+    }
+  }
+
+  &__list {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+
+    @media screen and (max-width: 768px) {
+      flex-direction: column;
+    }
+  }
+}
