+++ src/pages/Main/components/Insides/InsideItem/InsideItem.scss
@@ -1,0 +1,74 @@
+.insides-item {
+  display: flex;
+
+  @media screen and(max-width:768px) {
+    padding: 3% 0 2% 0;
+  }
+
+  &__wrapper {
+    display: inline-flex;
+    width: 100vw;
+
+    @media screen and(max-width:768px) {
+      flex-direction: column;
+    }
+  }
+
+  &__image {
+    width: 55%;
+
+    @media screen and(max-width:1190px) {
+      width: 50%;
+    }
+
+    @media screen and(max-width:768px) {
+      width: 100%;
+    }
+
+    >img {
+      width: 100%;
+    }
+  }
+
+  &__content {
+    width: 33%;
+    display: flex;
+    flex-direction: column;
+    justify-content: flex-end;
+    padding-left: 4%;
+
+    @media screen and(max-width:1190px) {
+      width: 29%;
+      padding-left: 2%;
+    }
+
+    @media screen and(max-width:768px) {
+      width: 100%;
+      padding: 1% 6.5%;
+    }
+  }
+
+  &__sub-title {
+    line-height: 20px;
+    letter-spacing: 1.5px;
+    font-weight: 600;
+    font-size: 10px;
+    color: #fff;
+  }
+
+  &__title {
+    font-size: 26px;
+    font-weight: 600;
+    color: #fff;
+
+    @media screen and(max-width:1190px) {
+      font-size: 25px;
+    }
+  }
+
+  &__text {
+    font-size: 15px;
+    line-height: 25px;
+    color: #fff;
+  }
+}
