+++ src/pages/Main/components/Insides/Insides.scss
@@ -1,0 +1,5 @@
+.insides__image-list {
+  display: flex;
+  padding: 2% 4.5%;
+  transition: opacity ease-out 0.6s 0.6s;
+}
