+++ src/pages/Main/components/Products/ProductsItem/ProductsItem.scss
@@ -1,0 +1,32 @@
+.products-item {
+  width: calc(100vw * 0.275);
+  padding-left: 2.4%;
+
+  @media screen and(max-width: 1190px) {
+    width: calc(100vw * 0.40);
+    padding-left: 2%;
+  }
+
+  @media screen and(max-width: 767px) {
+    width: calc(100vw * 0.8);
+    padding-left: 1%;
+  }
+
+  &__image {
+    >img {
+      width: 100%;
+      height: auto;
+    }
+  }
+
+  &__title {
+    line-height: 25px;
+    font-size: 15px;
+    text-decoration: underline;
+    color: #545454;
+
+    &:hover {
+      text-decoration: none;
+    }
+  }
+}
