+++ src/styles/common.scss
@@ -44,8 +44,8 @@
 
 
 @mixin button-large {
-  width: 135px;
-  height: 44px;
+  width: 132px;
+  height: 42px;
   border-radius: 21px;
   font-size: 14px;
   transition: all 0.3s ease-in-out;
